*Change: [KRF-00001] changed test service call to hook
